requirements.txt
fastapi
uvicorn
joblib
scikit-learn
openai
psycopg2-binary
python-dotenv
streamlit
requests
